add execution times of both retrieval at hybrid
