+++ tests/autorag/nodes/retrieval/test_run_retrieval_node.py
@@ -1,6 +1,7 @@
 import os.path
 import pathlib
 import shutil
+import tempfile
 
 import chromadb
 import pandas as pd
@@ -8,7 +9,7 @@ import pytest
 from llama_index import OpenAIEmbedding
 
 from autorag.nodes.retrieval import bm25, vectordb, hybrid_rrf
-from autorag.nodes.retrieval.run import run_retrieval_node
+from autorag.nodes.retrieval.run import run_retrieval_node, select_result_for_hybrid
 from autorag.nodes.retrieval.vectordb import vectordb_ingest
 from autorag.utils.util import load_summary_file
 
@@ -75,9 +76,88 @@ def test_run_retrieval_node(node_line_dir):
     assert summary_df['module_name'][0] == "bm25"
     assert summary_df['module_params'][0] == {'top_k': 4}
     assert summary_df['execution_time'][0] > 0
+    # assert average times
+    assert summary_df['execution_time'][0] + summary_df['execution_time'][1] == pytest.approx(
+        summary_df['execution_time'][2])
+
     # test the best file is saved properly
     best_filename = summary_df[summary_df['is_best'] == True]['filename'].values[0]
     best_path = os.path.join(node_line_dir, "retrieval", f'best_{best_filename}')
     assert os.path.exists(best_path)
     best_df = pd.read_parquet(best_path)
     assert all([expect_column in best_df.columns for expect_column in expect_columns])
+
+
+@pytest.fixture
+def pseudo_node_dir():
+    summary_df = pd.DataFrame({
+        'filename': ['0.parquet', '1.parquet', '2.parquet'],
+        'module_name': ['bm25', 'vectordb', 'vectordb'],
+        'module_params': [
+            {'top_k': 3},
+            {'top_k': 3, 'embedding_model': 'openai'},
+            {'top_k': 3, 'embedding_model': 'huggingface'},
+        ],
+        'execution_time': [1, 1, 1],
+        'retrieval_f1': [0.1, 0.2, 0.3],
+        'retrieval_recall': [0.2, 0.55, 0.5],
+    })
+    bm25_df = pd.DataFrame({
+        'query': ['query-1', 'query-2', 'query-3'],
+        'retrieved_ids': [['id-1', 'id-2', 'id-3'],
+                          ['id-1', 'id-2', 'id-3'],
+                          ['id-1', 'id-2', 'id-3']],
+        'retrieve_scores': [[0.1, 0.2, 0.3],
+                            [0.1, 0.2, 0.3],
+                            [0.1, 0.2, 0.3]],
+        'retrieval_f1': [0.05, 0.1, 0.15],
+        'retrieval_recall': [0.1, 0.275, 0.25],
+    })
+    vector_openai_df = pd.DataFrame({
+        'query': ['query-1', 'query-2', 'query-3'],
+        'retrieved_ids': [['id-4', 'id-5', 'id-6'],
+                          ['id-4', 'id-5', 'id-6'],
+                          ['id-4', 'id-5', 'id-6']],
+        'retrieve_scores': [[0.3, 0.4, 0.5],
+                            [0.3, 0.4, 0.5],
+                            [0.3, 0.4, 0.5]],
+        'retrieval_f1': [0.15, 0.2, 0.25],
+        'retrieval_recall': [0.3, 0.55, 0.5],
+    })
+    vector_huggingface_df = pd.DataFrame({
+        'query': ['query-1', 'query-2', 'query-3'],
+        'retrieved_ids': [['id-7', 'id-8', 'id-9'],
+                          ['id-7', 'id-8', 'id-9'],
+                          ['id-7', 'id-8', 'id-9']],
+        'retrieve_scores': [[0.5, 0.6, 0.7],
+                            [0.5, 0.6, 0.7],
+                            [0.5, 0.6, 0.7]],
+        'retrieval_f1': [0.25, 0.3, 0.35],
+        'retrieval_recall': [0.5, 0.825, 0.75],
+    })
+
+    with tempfile.TemporaryDirectory() as node_dir:
+        summary_df.to_csv(os.path.join(node_dir, "summary.csv"))
+        bm25_df.to_parquet(os.path.join(node_dir, "0.parquet"))
+        vector_openai_df.to_parquet(os.path.join(node_dir, "1.parquet"))
+        vector_huggingface_df.to_parquet(os.path.join(node_dir, "2.parquet"))
+        yield node_dir
+
+
+def test_select_result_for_hybrid(pseudo_node_dir):
+    ids, scores = select_result_for_hybrid(pseudo_node_dir, ("bm25", "vectordb"))
+    assert len(ids) == len(scores) == 2
+    assert len(ids[0]) == len(scores[0]) == 3
+    assert len(ids[1]) == len(scores[1]) == 3
+    assert ids[0] == [['id-1', 'id-2', 'id-3'],
+                      ['id-1', 'id-2', 'id-3'],
+                      ['id-1', 'id-2', 'id-3']]
+    assert scores[0] == [[0.1, 0.2, 0.3],
+                         [0.1, 0.2, 0.3],
+                         [0.1, 0.2, 0.3]]
+    assert ids[1] == [['id-7', 'id-8', 'id-9'],
+                      ['id-7', 'id-8', 'id-9'],
+                      ['id-7', 'id-8', 'id-9']]
+    assert scores[1] == [[0.5, 0.6, 0.7],
+                         [0.5, 0.6, 0.7],
+                         [0.5, 0.6, 0.7]]
